composites fix

+++ XP/xp03_conversation/conversation.py
@@ -240,6 +240,6 @@ def find_catalan_numbers(biggest_number: int) -> list:
     """
     catalan_numbers = [1]
     for i in range(1, biggest_number + 1):
-        number = int(((2 * (2 * i - 1)) / (i + 1)) * catalan_numbers[i - 1])
+        number = ((2 * (2 * i - 1)) / (i + 1)) * catalan_numbers[i - 1]
         catalan_numbers.append(number)
     return catalan_numbers
